+++ noodles-bgzf/src/async/writer/builder.rs
@@ -5,9 +5,8 @@ use tokio_util::codec::FramedWrite;
 
 use super::{Deflater, Writer};
 use crate::{
-    block,
     r#async::BlockCodec,
-    writer::{CompressionLevel, BGZF_EOF},
+    writer::{CompressionLevel, BGZF_EOF, DEFAULT_BUF_SIZE},
 };
 
 /// An async BGZF writer builder.
@@ -76,7 +75,7 @@ where
 
         Writer {
             sink: Deflater::new(FramedWrite::new(self.inner, BlockCodec)).buffer(worker_count),
-            buf: BytesMut::with_capacity(block::MAX_UNCOMPRESSED_DATA_LENGTH),
+            buf: BytesMut::with_capacity(DEFAULT_BUF_SIZE),
             eof_buf: Bytes::from_static(BGZF_EOF),
             compression_level: compression_level.into(),
         }